fix data type

+++ models/committee_user.js
@@ -22,17 +22,17 @@ module.exports = (sequelize, type) => {
       allowNull: false
     },
     role: {
-      type: type.DECIMAL(1, 0),
+      type: type.INTEGER,
       allowNull: false,
       defaultValue: 0
     },
     occupation: {
-      type: type.DECIMAL(1, 0),
+      type: type.INTEGER,
       allowNull: false,
       defaultValue: 0
     },
     disabled: {
-      type: type.DECIMAL(1, 0),
+      type: type.INTEGER,
       allowNull: false,
       defaultValue: 0
     }
