+++ models/vehicle.js
@@ -6,7 +6,7 @@ module.exports = (sequelize, type) => {
       autoIncrement: true
     },
     type: {
-      type: type.DECIMAL(1, 0),
+      type: type.INTEGER,
       allowNull: false
     },
     plate: {
